+++ internal/core/items.rs
@@ -29,16 +29,14 @@ use crate::item_rendering::{CachedRenderingData, RenderBorderRectangle};
 pub use crate::item_tree::ItemRc;
 use crate::layout::LayoutInfo;
 use crate::lengths::{
-    LogicalBorderRadius, LogicalLength, LogicalPoint, LogicalRect, LogicalSize, LogicalVector,
-    PointLengths, RectLengths,
+    LogicalBorderRadius, LogicalLength, LogicalSize, LogicalVector, PointLengths, RectLengths,
 };
 #[cfg(feature = "rtti")]
 use crate::rtti::*;
-use crate::window::{WindowAdapter, WindowAdapterRc, WindowInner};
-use crate::{Callback, Coord, Property, SharedString};
+use crate::window::{WindowAdapter, WindowAdapterRc};
+use crate::{Coord, Property, SharedString};
 use alloc::rc::Rc;
 use const_field_offset::FieldOffsets;
-use core::cell::Cell;
 use core::pin::Pin;
 use i_slint_core_macros::*;
 use vtable::*;
@@ -49,6 +47,8 @@ mod flickable;
 pub use flickable::*;
 mod text;
 pub use text::*;
+mod input_items;
+pub use input_items::*;
 mod image;
 pub use self::image::*;
 #[cfg(feature = "std")]
@@ -526,323 +526,8 @@ declare_item_vtable! {
     fn slint_get_BorderRectangleVTable() -> BorderRectangleVTable for BorderRectangle
 }
 
-/// The implementation of the `TouchArea` element
-#[repr(C)]
-#[derive(FieldOffsets, SlintElement, Default)]
-#[pin]
-pub struct TouchArea {
-    pub enabled: Property<bool>,
-    /// FIXME: We should annotate this as an "output" property.
-    pub pressed: Property<bool>,
-    pub has_hover: Property<bool>,
-    /// FIXME: there should be just one property for the point instead of two.
-    /// Could even be merged with pressed in a `Property<Option<Point>>` (of course, in the
-    /// implementation item only, for the compiler it would stay separate properties)
-    pub pressed_x: Property<LogicalLength>,
-    pub pressed_y: Property<LogicalLength>,
-    /// FIXME: should maybe be as parameter to the mouse event instead. Or at least just one property
-    pub mouse_x: Property<LogicalLength>,
-    pub mouse_y: Property<LogicalLength>,
-    pub mouse_cursor: Property<MouseCursor>,
-    pub clicked: Callback<VoidArg>,
-    pub double_clicked: Callback<VoidArg>,
-    pub moved: Callback<VoidArg>,
-    pub pointer_event: Callback<PointerEventArg>,
-    pub scroll_event: Callback<PointerScrollEventArg, EventResult>,
-    /// FIXME: remove this
-    pub cached_rendering_data: CachedRenderingData,
-    /// true when we are currently grabbing the mouse
-    grabbed: Cell<bool>,
-}
-
-impl Item for TouchArea {
-    fn init(self: Pin<&Self>, _self_rc: &ItemRc) {}
-
-    fn layout_info(
-        self: Pin<&Self>,
-        _orientation: Orientation,
-        _window_adapter: &Rc<dyn WindowAdapter>,
-    ) -> LayoutInfo {
-        LayoutInfo { stretch: 1., ..LayoutInfo::default() }
-    }
-
-    fn input_event_filter_before_children(
-        self: Pin<&Self>,
-        event: MouseEvent,
-        window_adapter: &Rc<dyn WindowAdapter>,
-        _self_rc: &ItemRc,
-    ) -> InputEventFilterResult {
-        if !self.enabled() {
-            return InputEventFilterResult::ForwardAndIgnore;
-        }
-        if let Some(pos) = event.position() {
-            Self::FIELD_OFFSETS.mouse_x.apply_pin(self).set(pos.x_length());
-            Self::FIELD_OFFSETS.mouse_y.apply_pin(self).set(pos.y_length());
-        }
-        let hovering = !matches!(event, MouseEvent::Exit);
-        Self::FIELD_OFFSETS.has_hover.apply_pin(self).set(hovering);
-        if hovering {
-            if let Some(x) = window_adapter.internal(crate::InternalToken) {
-                x.set_mouse_cursor(self.mouse_cursor());
-            }
-        }
-        InputEventFilterResult::ForwardAndInterceptGrab
-    }
-
-    fn input_event(
-        self: Pin<&Self>,
-        event: MouseEvent,
-        window_adapter: &Rc<dyn WindowAdapter>,
-        self_rc: &ItemRc,
-    ) -> InputEventResult {
-        if matches!(event, MouseEvent::Exit) {
-            Self::FIELD_OFFSETS.has_hover.apply_pin(self).set(false);
-            if let Some(x) = window_adapter.internal(crate::InternalToken) {
-                x.set_mouse_cursor(MouseCursor::Default);
-            }
-        }
-        if !self.enabled() {
-            return InputEventResult::EventIgnored;
-        }
-
-        match event {
-            MouseEvent::Pressed { position, button, .. } => {
-                self.grabbed.set(true);
-                if button == PointerEventButton::Left {
-                    Self::FIELD_OFFSETS.pressed_x.apply_pin(self).set(position.x_length());
-                    Self::FIELD_OFFSETS.pressed_y.apply_pin(self).set(position.y_length());
-                    Self::FIELD_OFFSETS.pressed.apply_pin(self).set(true);
-                }
-                Self::FIELD_OFFSETS.pointer_event.apply_pin(self).call(&(PointerEvent {
-                    button,
-                    kind: PointerEventKind::Down,
-                    modifiers: window_adapter.window().0.modifiers.get().into(),
-                },));
-
-                InputEventResult::GrabMouse
-            }
-            MouseEvent::Exit => {
-                Self::FIELD_OFFSETS.pressed.apply_pin(self).set(false);
-                if self.grabbed.replace(false) {
-                    Self::FIELD_OFFSETS.pointer_event.apply_pin(self).call(&(PointerEvent {
-                        button: PointerEventButton::Other,
-                        kind: PointerEventKind::Cancel,
-                        modifiers: window_adapter.window().0.modifiers.get().into(),
-                    },));
-                }
-
-                InputEventResult::EventAccepted
-            }
-
-            MouseEvent::Released { button, position, click_count } => {
-                let geometry = self_rc.geometry();
-                if button == PointerEventButton::Left
-                    && LogicalRect::new(LogicalPoint::default(), geometry.size).contains(position)
-                    && self.pressed()
-                {
-                    Self::FIELD_OFFSETS.clicked.apply_pin(self).call(&());
-                    if (click_count % 2) == 1 {
-                        Self::FIELD_OFFSETS.double_clicked.apply_pin(self).call(&())
-                    }
-                }
-
-                self.grabbed.set(false);
-                if button == PointerEventButton::Left {
-                    Self::FIELD_OFFSETS.pressed.apply_pin(self).set(false);
-                }
-                Self::FIELD_OFFSETS.pointer_event.apply_pin(self).call(&(PointerEvent {
-                    button,
-                    kind: PointerEventKind::Up,
-                    modifiers: window_adapter.window().0.modifiers.get().into(),
-                },));
-
-                InputEventResult::EventAccepted
-            }
-            MouseEvent::Moved { .. } => {
-                Self::FIELD_OFFSETS.pointer_event.apply_pin(self).call(&(PointerEvent {
-                    button: PointerEventButton::Other,
-                    kind: PointerEventKind::Move,
-                    modifiers: window_adapter.window().0.modifiers.get().into(),
-                },));
-                return if self.grabbed.get() {
-                    Self::FIELD_OFFSETS.moved.apply_pin(self).call(&());
-                    InputEventResult::GrabMouse
-                } else {
-                    InputEventResult::EventAccepted
-                };
-            }
-            MouseEvent::Wheel { delta_x, delta_y, .. } => {
-                let modifiers = window_adapter.window().0.modifiers.get().into();
-                let r = Self::FIELD_OFFSETS
-                    .scroll_event
-                    .apply_pin(self)
-                    .call(&(PointerScrollEvent { delta_x, delta_y, modifiers },));
-                if self.grabbed.get() {
-                    InputEventResult::GrabMouse
-                } else {
-                    match r {
-                        EventResult::Reject => {
-                            // We are ignoring the event, so we will be removed from the item_stack,
-                            // therefore we must remove the has_hover flag as there might be a scroll under us.
-                            // It will be put back later.
-                            Self::FIELD_OFFSETS.has_hover.apply_pin(self).set(false);
-                            InputEventResult::EventIgnored
-                        }
-                        EventResult::Accept => InputEventResult::EventAccepted,
-                    }
-                }
-            }
-        }
-    }
-
-    fn key_event(
-        self: Pin<&Self>,
-        _: &KeyEvent,
-        _window_adapter: &Rc<dyn WindowAdapter>,
-        _self_rc: &ItemRc,
-    ) -> KeyEventResult {
-        KeyEventResult::EventIgnored
-    }
-
-    fn focus_event(
-        self: Pin<&Self>,
-        _: &FocusEvent,
-        _window_adapter: &Rc<dyn WindowAdapter>,
-        _self_rc: &ItemRc,
-    ) -> FocusEventResult {
-        FocusEventResult::FocusIgnored
-    }
-
-    fn render(
-        self: Pin<&Self>,
-        _backend: &mut ItemRendererRef,
-        _self_rc: &ItemRc,
-        _size: LogicalSize,
-    ) -> RenderingResult {
-        RenderingResult::ContinueRenderingChildren
-    }
-}
-
-impl ItemConsts for TouchArea {
-    const cached_rendering_data_offset: const_field_offset::FieldOffset<
-        TouchArea,
-        CachedRenderingData,
-    > = TouchArea::FIELD_OFFSETS.cached_rendering_data.as_unpinned_projection();
-}
-
 declare_item_vtable! {
     fn slint_get_TouchAreaVTable() -> TouchAreaVTable for TouchArea
-}
-
-/// A runtime item that exposes key
-#[repr(C)]
-#[derive(FieldOffsets, Default, SlintElement)]
-#[pin]
-pub struct FocusScope {
-    pub enabled: Property<bool>,
-    pub has_focus: Property<bool>,
-    pub key_pressed: Callback<KeyEventArg, EventResult>,
-    pub key_released: Callback<KeyEventArg, EventResult>,
-    pub focus_changed_event: Callback<VoidArg>,
-    /// FIXME: remove this
-    pub cached_rendering_data: CachedRenderingData,
-}
-
-impl Item for FocusScope {
-    fn init(self: Pin<&Self>, _self_rc: &ItemRc) {}
-
-    fn layout_info(
-        self: Pin<&Self>,
-        _orientation: Orientation,
-        _window_adapter: &Rc<dyn WindowAdapter>,
-    ) -> LayoutInfo {
-        LayoutInfo { stretch: 1., ..LayoutInfo::default() }
-    }
-
-    fn input_event_filter_before_children(
-        self: Pin<&Self>,
-        _: MouseEvent,
-        _window_adapter: &Rc<dyn WindowAdapter>,
-        _self_rc: &ItemRc,
-    ) -> InputEventFilterResult {
-        InputEventFilterResult::ForwardEvent
-    }
-
-    fn input_event(
-        self: Pin<&Self>,
-        event: MouseEvent,
-        window_adapter: &Rc<dyn WindowAdapter>,
-        self_rc: &ItemRc,
-    ) -> InputEventResult {
-        if self.enabled() && matches!(event, MouseEvent::Pressed { .. }) && !self.has_focus() {
-            WindowInner::from_pub(window_adapter.window()).set_focus_item(self_rc, true);
-            InputEventResult::EventAccepted
-        } else {
-            InputEventResult::EventIgnored
-        }
-    }
-
-    fn key_event(
-        self: Pin<&Self>,
-        event: &KeyEvent,
-        _window_adapter: &Rc<dyn WindowAdapter>,
-        _self_rc: &ItemRc,
-    ) -> KeyEventResult {
-        let r = match event.event_type {
-            KeyEventType::KeyPressed => {
-                Self::FIELD_OFFSETS.key_pressed.apply_pin(self).call(&(event.clone(),))
-            }
-            KeyEventType::KeyReleased => {
-                Self::FIELD_OFFSETS.key_released.apply_pin(self).call(&(event.clone(),))
-            }
-            KeyEventType::UpdateComposition | KeyEventType::CommitComposition => {
-                EventResult::Reject
-            }
-        };
-        match r {
-            EventResult::Accept => KeyEventResult::EventAccepted,
-            EventResult::Reject => KeyEventResult::EventIgnored,
-        }
-    }
-
-    fn focus_event(
-        self: Pin<&Self>,
-        event: &FocusEvent,
-        _window_adapter: &Rc<dyn WindowAdapter>,
-        _self_rc: &ItemRc,
-    ) -> FocusEventResult {
-        if !self.enabled() {
-            return FocusEventResult::FocusIgnored;
-        }
-
-        match event {
-            FocusEvent::FocusIn | FocusEvent::WindowReceivedFocus => {
-                self.has_focus.set(true);
-                Self::FIELD_OFFSETS.focus_changed_event.apply_pin(self).call(&());
-            }
-            FocusEvent::FocusOut | FocusEvent::WindowLostFocus => {
-                self.has_focus.set(false);
-                Self::FIELD_OFFSETS.focus_changed_event.apply_pin(self).call(&());
-            }
-        }
-        FocusEventResult::FocusAccepted
-    }
-
-    fn render(
-        self: Pin<&Self>,
-        _backend: &mut ItemRendererRef,
-        _self_rc: &ItemRc,
-        _size: LogicalSize,
-    ) -> RenderingResult {
-        RenderingResult::ContinueRenderingChildren
-    }
-}
-
-impl ItemConsts for FocusScope {
-    const cached_rendering_data_offset: const_field_offset::FieldOffset<
-        FocusScope,
-        CachedRenderingData,
-    > = FocusScope::FIELD_OFFSETS.cached_rendering_data.as_unpinned_projection();
 }
 
 declare_item_vtable! {
@@ -1529,7 +1214,7 @@ i_slint_common::for_each_builtin_structs!(declare_builtin_structs);
 pub unsafe extern "C" fn slint_item_absolute_position(
     self_component: &vtable::VRc<crate::item_tree::ItemTreeVTable>,
     self_index: u32,
-) -> LogicalPoint {
+) -> crate::lengths::LogicalPoint {
     let self_rc = ItemRc::new(self_component.clone(), self_index);
     self_rc.map_to_window(Default::default())
 }
